aider: Update task names to include repository name as a variable.
# Aider chat conversation:

USER: Is there a way to parameterize the task names so that output would include 'Add APT repository for {{repo_name}}'?

ASSISTANT: To parameterize the task names to include the repository name, I will update the `add-repo.yml` file to use a variable `repo_name` in the task names. Then, I will update each playbook that includes `add-repo.yml` to pass the `repo_name` variable.

Here are the diffs for the changes:

```diff
--- vm-setup/add-repo.yml
+++ vm-setup/add-repo.yml
@@ ... @@
 - name: Add APT Repository
   hosts: localhost
   become: yes
   tasks:
-    - name: Install Repository Public Key
+    - name: "Install Repository Public Key for {{ repo_name }}"
       apt_key:
         url: "{{ key_url }}"
         state: present

-    - name: Add Repository
+    - name: "Add APT repository for {{ repo_name }}"
       apt_repository:
         repo: "{{ repo_url }}"
         filename: "{{ filename }}"
         state: present
```

Now, I will update each playbook to pass the `repo_name` variable:

```diff
--- vm-setup/repo-hashicorp.yml
+++ vm-setup/repo-hashicorp.yml
@@ ... @@
   vars:
     key_url: https://apt.releases.hashicorp.com/gpg
     repo_url: deb https://apt.releases.hashicorp.com focal main
     filename: hashicorp
+    repo_name: Hashicorp

   tasks:
     - include_tasks: add-repo.yml
```

```diff
--- vm-setup/repo-google-cloud.yml
+++ vm-setup/repo-google-cloud.yml
@@ ... @@
   vars:
     key_url: https://packages.cloud.google.com/apt/doc/apt-key.gpg
     repo_url: deb https://packages.cloud.google.com/apt cloud-sdk main
     filename: google-cloud
+    repo_name: Google Cloud

   tasks:
     - include_tasks: add-repo.yml
```

```diff
--- vm-setup/repo-microsoft.yml
+++ vm-setup/repo-microsoft.yml
@@ ... @@
   vars:
     key_url: https://packages.microsoft.com/keys/microsoft.asc
     repo_url: deb https://packages.microsoft.com/ubuntu/20.04/prod focal main
     filename: microsoft
+    repo_name: Microsoft

   tasks:
     - include_tasks: add-repo.yml
```

These changes will ensure that the output of the tasks will include the name of the repository being added, making the output more descriptive and easier to understand.