aider: Modify tasks/main.yml to error out with a message and a list of available repositories if a repository is not in the defaults.
# Aider chat conversation:

USER: Modify tasks/main.yml to error out with a message and a list of available repositories if a repository is not in the defaults.

ASSISTANT: To modify `tasks/main.yml` to error out with a message and a list of available repositories if a repository is not in the defaults, here is the diff:

```diff
--- vm-setup/roles/repo_add/tasks/main.yml
+++ vm-setup/roles/repo_add/tasks/main.yml
@@ -1,11 +1,23 @@
 ---
-- name: Install specified repositories or default to all
+- name: Check if all requested repositories are defined in defaults
+  vars:
+    all_repos: "{{ repos.keys() }}"
+    requested_repos: "{{ repo_list | default(all_repos) }}"
+  fail:
+    msg: "The following requested repositories are not defined in defaults: {{ requested_repos | difference(all_repos) | join(', ') }}. Available repositories are: {{ all_repos | join(', ') }}"
+  when: requested_repos | difference(all_repos) | length > 0
+
+- name: Install requested repositories or default to all
   vars:
-    repo_subset: "{{ repo_list | default(repos.keys()) }}"
+    repo_subset: "{{ requested_repos }}"

   loop: "{{ repo_subset }}"
   loop_control:
     loop_var: repo_name

   include_tasks: install_repo.yml

   vars:
     key_url: "{{ repos[repo_name].key_url }}"
     repo_url: "{{ repos[repo_name].repo_url }}"
     filename: "{{ repos[repo_name].filename }}"
     repo_name: "{{ repos[repo_name].repo_name }}"
```

This change adds a task at the beginning of `tasks/main.yml` that checks if all repositories listed in `repo_list` (or all repositories by default if `repo_list` is not defined) are present in the `repos` hash from `defaults/main.yml`. If any requested repository is not defined, it fails with an error message listing the undefined repositories and all available repositories. If all requested repositories are defined, it proceeds to install them as before.