aider: Split fail message in tasks/main.yml to list undefined repos on one line and available repos on the next line.
# Aider chat conversation:

USER: Can the fail message in tasks/main.yml be split so that undefined repos are on one line and available repos are on the next line?

ASSISTANT: To split the fail message in `tasks/main.yml` so that undefined repositories are listed on one line and available repositories are on the next line, here is the diff:

```diff
--- vm-setup/roles/repo_add/tasks/main.yml
+++ vm-setup/roles/repo_add/tasks/main.yml
@@ -5,7 +5,8 @@
   vars:
     all_repos: "{{ repos.keys() }}"
     requested_repos: "{{ repo_list | default(all_repos) }}"
-
+  fail:
+    msg: >
+      These repositories are not defined: {{ requested_repos | difference(all_repos) | join(', ') }}.
+      Available repositories: {{ all_repos | join(', ') }}
   when: requested_repos | difference(all_repos) | length > 0

 - name: Install requested repositories
```

This change modifies the fail message to use the folded style block scalar (`>`) in YAML, which allows for newlines to be preserved in the output. The message now has the list of undefined repositories on one line and the list of available repositories on the next line.