aider: Rename repo_add to add_repo.
# Aider chat conversation:

USER: Rename repo_add to add_repo.

ASSISTANT: To rename the role `repo_add` to `add_repo`, you will need to update the references in the playbook and the directory structure. Here are the diffs for the necessary changes:

```diff
--- vm-setup/sweetums.yml
+++ vm-setup/sweetums.yml
@@ ... @@
   roles:
-    - role: repo_add
+    - role: add_repo
       vars:
         repo_list:
           - hashicorp
#          - badrepo
```

Please note that you will also need to rename the directory from `vm-setup/roles/repo_add` to `vm-setup/roles/add_repo`. This action cannot be represented as a diff and will need to be performed manually or with a command such as `mv vm-setup/roles/repo_add vm-setup/roles/add_repo`.