aider: Corrected the relative path to the `test_add_repo.yml` playbook from the `converge.yml` file.
# Aider chat conversation:

USER: I ran this command:

 cd ansible/harleypig/env_setup/extensions ; molecule converge # What is causing this error?

And got this output:

[34mINFO    [0m default scenario test matrix: dependency, create, prepare, converge
[34mINFO    [0m Performing prerun with [33mrole_name_check[0m=[1;36m0[0m[33m...[0m
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mdependency[0m
[31mWARNING [0m Skipping, missing the requirements file.
[31mWARNING [0m Skipping, missing the requirements file.
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mcreate[0m
[31mWARNING [0m Skipping, instances already created.
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mprepare[0m
[31mWARNING [0m Skipping, prepare playbook not configured.
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mconverge[0m
[31mERROR! Unable to retrieve file contents[0m
[31mCould not find or access '/home/harleypig/projects/dotfiles/ansible/harleypig/env_setup/extensions/molecule/default/harleypig.env_setup.test_add_repo' on the Ansible Controller.[0m
[31mIf you are using a module and expect the file to exist on the remote, see the remote_src option[0m
[1;31mCRITICAL[0m Ansible return code was 1, command was: [2mansible-playbook --inventory /home/harleypig/.cache/molecule/extensions/default/inventory --skip-tags molecule-notest,notest /home/harleypig/projects/dotfiles/ansible/harleypig/env_setup/extensions/molecule/default/converge.yml[0m

ASSISTANT: The error you're encountering is due to the Ansible playbook `harleypig.env_setup.test_add_repo` not being found at the expected location. The error message indicates that Ansible is looking for the file in the directory `/home/harleypig/projects/dotfiles/ansible/harleypig/env_setup/extensions/molecule/default/`, but it cannot find it.

The issue seems to be related to the incorrect path specified in the `converge.yml` file. The `import_playbook` directive is looking for `harleypig.env_setup.test_add_repo`, but it should be looking for `test_add_repo.yml` which is the correct filename.

Here is the diff to fix the path in the `converge.yml` file:

```diff
--- ansible/harleypig/env_setup/extensions/molecule/default/converge.yml
+++ ansible/harleypig/env_setup/extensions/molecule/default/converge.yml
@@ ... @@
-  ansible.builtin.import_playbook: harleypig.env_setup.test_add_repo
+  ansible.builtin.import_playbook: ../../playbooks/test_add_repo.yml
```

This change corrects the relative path to the `test_add_repo.yml` playbook from the `converge.yml` file. After applying this patch, you should be able to run the `molecule converge` command without encountering the previous error.