aider: The platform name in the molecule yaml file has been updated to use the `ubuntu:20.04` Docker image for testing.
# Aider chat conversation:

USER: I am having network failure (a name resolution error). However, I am able to run wget https://packages.cloud.google.com/apt/doc/apt-key.gpg and retreive it from the command line. Please wait to respond until my next prompt where I will include the full error.

ASSISTANT: Understood. I'll wait for your next prompt with the full error message before proceeding with any changes or suggestions.

USER: I ran this command:

 cd ansible/harleypig/env_setup/roles/add_repo ; molecule converge # Molecule is using docker, but if the other two repos are working why is google-cloud failing?

And got this output:

[34mINFO    [0m default scenario test matrix: dependency, create, prepare, converge
[34mINFO    [0m Performing prerun with [33mrole_name_check[0m=[1;36m0[0m[33m...[0m
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mdependency[0m
[31mWARNING [0m Skipping, missing the requirements file.
[31mWARNING [0m Skipping, missing the requirements file.
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mcreate[0m
[31mWARNING [0m Skipping, instances already created.
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mprepare[0m
[31mWARNING [0m Skipping, prepare playbook not configured.
[34mINFO    [0m [2;36mRunning [0m[2;32mdefault[0m[2;36m > [0m[2;32mconverge[0m

PLAY [Converge] ****************************************************************

TASK [Test add_repo role (converge)] *******************************************

TASK [harleypig.env_setup.add_repo : Validate requested repositories] **********
[36mskipping: [instance][0m

TASK [harleypig.env_setup.add_repo : Install requested repositories] ***********
[36mincluded: /home/harleypig/.ansible/collections/ansible_collections/harleypig/env_setup/roles/add_repo/tasks/install_repo.yml for instance => (item=google-cloud)[0m
[36mincluded: /home/harleypig/.ansible/collections/ansible_collections/harleypig/env_setup/roles/add_repo/tasks/install_repo.yml for instance => (item=hashicorp)[0m
[36mincluded: /home/harleypig/.ansible/collections/ansible_collections/harleypig/env_setup/roles/add_repo/tasks/install_repo.yml for instance => (item=microsoft)[0m

TASK [harleypig.env_setup.add_repo : Validate required variables for Google Cloud] ***
[32mok: [instance] => {[0m
[32m    "changed": false,[0m
[32m    "msg": "All assertions passed"[0m
[32m}[0m

TASK [harleypig.env_setup.add_repo : Install Repository Public Key for Google Cloud] ***
[1;31mfatal: [instance]: UNREACHABLE! => {"changed": false, "msg": "Failed to connect to the host via ssh: ssh: Could not resolve hostname instance: Temporary failure in name resolution", "unreachable": true}[0m

PLAY RECAP *********************************************************************
[31minstance[0m                   : [32mok=4   [0m changed=0    [1;31munreachable=1   [0m failed=0    [36mskipped=1   [0m rescued=0    ignored=0

[1;31mCRITICAL[0m Ansible return code was 4, command was: [2mansible-playbook --inventory /home/harleypig/.cache/molecule/add_repo/default/inventory --skip-tags molecule-notest,notest /home/harleypig/projects/dotfiles/ansible/harleypig/env_setup/roles/add_repo/molecule/default/converge.yml[0m

ASSISTANT: The error message indicates that the Ansible task is trying to connect to a host named "instance" via SSH and is unable to resolve the hostname, leading to a name resolution error. This suggests that there might be an issue with the configuration of the molecule test environment, specifically with the definition of the test instance.

The `molecule.yml` file within the `molecule/default` directory is responsible for defining the platforms (test instances) that Molecule will use to test the role. It seems that the `instance` is not properly configured, which could be causing the name resolution error.

To address this issue, you would typically need to ensure that the `molecule.yml` file contains the correct configuration for the Docker container that Molecule will use for testing. This includes specifying the Docker image and any other necessary settings.

However, since I am only able to provide diffs for changes, I would need to know what specific changes you would like to make to the `molecule.yml` file or any other files to help resolve the network failure issue. If you have a specific configuration in mind or need assistance determining the correct configuration, please let me know, and I can provide the appropriate diffs.

USER: in the env_setup role, in the molecule yaml file, what should the platform name be? Is this a docker image name (since I'm using docker)?

ASSISTANT: In the `molecule.yml` file, the platform name is an identifier for the test instance, and it does not necessarily have to be a Docker image name. However, you do need to specify the Docker image you want to use for the test environment under the `image` key within the platform's definition.

Since you are using Docker with Molecule, you should specify a Docker image that Molecule will use to create the test container. This image should be one that is compatible with the role you are testing. For example, if your role is meant to run on an Ubuntu system, you might use the `ubuntu:20.04` image.

Here is an example diff to update the `molecule.yml` file to use the `ubuntu:20.04` Docker image:

```diff
--- ansible/harleypig/env_setup/roles/add_repo/molecule/default/molecule.yml
+++ ansible/harleypig/env_setup/roles/add_repo/molecule/default/molecule.yml
@@ ... @@
 platforms:
   - name: instance
+    image: ubuntu:20.04
+    pre_build_image: true
```

This diff adds the `image` key with the value `ubuntu:20.04` and the `pre_build_image` key with the value `true` to the platform definition. The `pre_build_image` key tells Molecule to use the image as-is without trying to build it from a Dockerfile.

Please replace `ubuntu:20.04` with the Docker image that is appropriate for your role's testing needs. If you have a specific image in mind, let me know, and I can update the diff accordingly.