aider: Modify assert task to dynamically report undefined variables in install_repo.yml.
# Aider chat conversation:

USER: in add_repo/tasks/install_repo.yml is there a way to output the failed variable?

ASSISTANT: To output the name of the failed variable in `add_repo/tasks/install_repo.yml`, you can modify the `assert` task to include a `fail_msg` that dynamically reports which variable is undefined. Here's the diff for the changes:

```diff
--- ansible/harleypig/env_setup/roles/add_repo/tasks/install_repo.yml
+++ ansible/harleypig/env_setup/roles/add_repo/tasks/install_repo.yml
@@ ... @@
 - name: "Validate required variables for {{ repo_name }}"
   assert:
     that:
       - key_url is defined
       - repo_url is defined
       - filename is defined
       - repo_name is defined
-    fail_msg: "Required variable is not defined for {{ repo_name }}."
+    fail_msg: "{{ item }} is not defined for {{ repo_name }}"
+  with_items:
+    - key_url
+    - repo_url
+    - filename
+    - repo_name
+  loop_control:
+    label: "{{ item }}"
```

This change will loop over each variable and check if it's defined. If any of them is not, the `fail_msg` will output the name of the undefined variable along with the repository name.